all influence tables updated for 0 contrast exp
Tables now check to see if an experiment has contrast data (for single
trials, stored in stim_vis_dir second column). If it does, it takes all
trials with 0 contrast and assigns them a direction of '0', since this
is unused in my experiments (I use 360 instead). Analysis then treats
all of these 0-contrast trials as a single trial-type.